+++ src/vram/bitmap.rs
